Added validation before submitting data

+++ ProjectManagementAPI/Persistance/Repositories/ProjectRepository.cs
@@ -1,28 +1,32 @@
-﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.EntityFrameworkCore;
 using ProjectManagementAPI.Domain.Repositories;
 using ProjectManagementAPI.Models;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Task = System.Threading.Tasks.Task;
 
 namespace ProjectManagementAPI.Persistance.Repositories
 {
     public class ProjectRepository : BaseRepository, IProjectRepository
     {
-        public ProjectRepository(AppDbContext context) : base(context){}
+        public ProjectRepository(AppDbContext context) : base(context) { }
 
-        public async Task CreateProject(Project project)
+        public async Task<bool> CreateProject(Project project)
         {
             _context.Projects.Add(project);
             await _context.SaveChangesAsync();
+            return true;
         }
 
-        public async Task DeleteProject(long id)
+        public async Task<bool> DeleteProject(long id)
         {
             var project = _context.Projects.Find(id);
+            if (project == null)
+            {
+                return false;
+            }
             _context.Projects.Remove(project);
             await _context.SaveChangesAsync();
+            return true;
         }
 
         public async Task<Project> GetProject(long id)
@@ -36,14 +40,22 @@ namespace ProjectManagementAPI.Persistance.Repositories
 
         public async Task<IEnumerable<Project>> GetProjects()
         {
-            return await _context.Projects.Include(p => p.Tasks).Include(p => p.Owner).ToListAsync();
+            return await _context.Projects
+                .Include(p => p.Tasks)
+                .Include(p => p.Owner)
+                .ToListAsync();
         }
 
-        public async Task UpdateProject(long id, Project project)
+        public async Task<bool> UpdateProject(long id, Project project)
         {
-            project.Id = id;
+            var foundProject = await _context.Projects.FindAsync(id);
+            if (foundProject == null)
+            {
+                return false;
+            }
             _context.Entry(project).State = EntityState.Modified;
             await _context.SaveChangesAsync();
+            return true;
         }
     }
 }
